fully refactored ris_error
