+++ tests/suite/main.rs
@@ -1,7 +1,7 @@
 pub mod ris_asset;
 pub mod ris_data;
+pub mod ris_file;
 pub mod ris_input;
 pub mod ris_jobs;
 pub mod ris_math;
 pub mod ris_rng;
-pub mod ris_util;
