+++ tests/suite/ris_file/fallback_file_append.rs
@@ -8,7 +8,7 @@ use chrono::DateTime;
 use chrono::Duration;
 use chrono::Local;
 
-use ris_util::fallback_file::FallbackFileAppend;
+use ris_file::fallback_file::FallbackFileAppend;
 
 #[test]
 fn should_create_directories() {
